+++ CHANGELOG.md
@@ -73,3 +73,6 @@
 
 * train_manager_apso_f1.py
 * eval_manager.py
+* train_manager_f1.py
+* eval_experts_meld.py
+* Folder fine_tuning